+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix dependency to Adafruit MCP23017 Arduino Library in library.properties.
+
 ## [0.4.0]
 
 ### Added