+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.4.1] - 2026-07-16
+
 ### Fixed
 
 - Fix dependency to Adafruit MCP23017 Arduino Library in library.properties.
@@ -85,7 +87,8 @@ All notable changes to this project will be documented in this file.
 
 - First public release
 
-[unreleased]: https://github.com/soosp/DigitalOutput/compare/0.4.0...HEAD
+[unreleased]: https://github.com/soosp/DigitalOutput/compare/0.4.1...HEAD
+[0.4.1]: https://github.com/soosp/DigitalOutput/compare/0.4.0...0.4.1
 [0.4.0]: https://github.com/soosp/DigitalOutput/compare/0.3.1...0.4.0
 [0.3.1]: https://github.com/soosp/DigitalOutput/compare/0.3.0...0.3.1
 [0.3.0]: https://github.com/soosp/DigitalOutput/compare/0.2.0...0.3.0